Добаляет файлы data.js и util.js и распределяет код исходя из назначения файла
